Refactored styled table rows and cells
